@@ -8,7 +8,7 @@ description = "TODO"
 version = "0.1.2"
 readme = "README.md"
 license = { file = "LICENSE" }
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 authors = [
     { name = "Point72", email = "OpenSource@Point72.com" },
 ]
