@@ -16,9 +16,10 @@ Notes:
 
 import inspect
 from typing import Any, Callable, Dict, List, Union
+from pydantic import GetJsonSchemaHandler
+from pydantic.json_schema import JsonSchemaValue
+from pydantic_core import core_schema
 
-import pydantic
-from packaging import version
 
 __all__ = ("auto", "Enum", "make_enum")
 
@@ -54,106 +55,52 @@ except ImportError:
         _CSP_ENUM = False
 
 
-if version.parse(pydantic.__version__) < version.parse("2"):
+class Enum(BaseEnum):
+    """Hybrid enum type based on csp.Enum if csp is available, otherwise enum.Enum"""
 
-    class Enum(BaseEnum):
-        """Hybrid enum type based on csp.Enum if csp is available, otherwise enum.Enum"""
+    # NOTE: we overload the default numbering
+    # behavior to start at 0 and match to csp
+    def _generate_next_value_(name, start, count, last_values):
+        return count
 
-        # NOTE: we overload the default numbering
-        # behavior to start at 0 and match to csp
-        def _generate_next_value_(name, start, count, last_values):
-            return count
+    @classmethod
+    def validate(cls, v) -> "Enum":
+        if isinstance(v, cls):
+            return v
+        elif isinstance(v, str):
+            return cls[v]
+        elif isinstance(v, int):
+            return cls(v)
+        raise ValueError(f"Cannot convert value to enum: {v}")
 
-        # TODO mainline to csp
-        # Add this to make our enums play super nicely with pydantic
-        @classmethod
-        def __get_validators__(cls):
-            yield cls.validate
+    @staticmethod
+    def _serialize(value: "Enum") -> str:
+        return value.name
 
-        @classmethod
-        def validate(cls, v, field=None):
-            if isinstance(v, cls):
-                return v
-            elif isinstance(v, str):
-                return cls[v]
-            elif isinstance(v, int):
-                return cls(v)
-            raise TypeError(f"Cannot convert value to enum: {v}")
+    @classmethod
+    def __get_pydantic_json_schema__(cls, _core_schema: core_schema.CoreSchema, handler: GetJsonSchemaHandler) -> JsonSchemaValue:
+        field_schema = handler(core_schema.str_schema())
+        field_schema.update(
+            type="string",
+            title=cls.__name__,
+            description=cls.__doc__ or "An enumeration of {}".format(cls.__name__),
+            enum=list(cls.__members__.keys()),
+        )
+        return field_schema
 
-        @classmethod
-        def __modify_schema__(cls, field_schema):
-            # Necessary class method for providing a JSON Schema
-            # https://github.com/pydantic/pydantic/blob/0a6e41386623d48b3ccd572e2b6c3e7238ee7002/pydantic/schema.py#L654
-            field_schema.update(
-                type="string",
-                title=cls.__name__,
-                description=cls.__doc__ or "An enumeration of {}".format(cls.__name__),
-                enum=list(cls.__members__.keys()),
-            )
-
-    from pydantic.json import ENCODERS_BY_TYPE
-
-    # Use `name` of enum when json encoding
-    ENCODERS_BY_TYPE[Enum] = lambda x: x.name
-    ENCODERS_BY_TYPE[BaseEnum] = ENCODERS_BY_TYPE[Enum]
-
-    if _CSP_ENUM:
-        # Workaround for how pydantic checks type
-        from csp.impl.enum import EnumMeta
-
-        ENCODERS_BY_TYPE[EnumMeta] = ENCODERS_BY_TYPE[Enum]
-
-else:
-    from pydantic import GetJsonSchemaHandler
-    from pydantic.json_schema import JsonSchemaValue
-    from pydantic_core import core_schema
-
-    class Enum(BaseEnum):
-        """Hybrid enum type based on csp.Enum if csp is available, otherwise enum.Enum"""
-
-        # NOTE: we overload the default numbering
-        # behavior to start at 0 and match to csp
-        def _generate_next_value_(name, start, count, last_values):
-            return count
-
-        @classmethod
-        def validate(cls, v) -> "Enum":
-            if isinstance(v, cls):
-                return v
-            elif isinstance(v, str):
-                return cls[v]
-            elif isinstance(v, int):
-                return cls(v)
-            raise ValueError(f"Cannot convert value to enum: {v}")
-
-        @staticmethod
-        def _serialize(value: "Enum") -> str:
-            return value.name
-
-        @classmethod
-        def __get_pydantic_json_schema__(cls, _core_schema: core_schema.CoreSchema, handler: GetJsonSchemaHandler) -> JsonSchemaValue:
-            field_schema = handler(core_schema.str_schema())
-            field_schema.update(
-                type="string",
-                title=cls.__name__,
-                description=cls.__doc__ or "An enumeration of {}".format(cls.__name__),
-                enum=list(cls.__members__.keys()),
-            )
-            return field_schema
-
-        @classmethod
-        def __get_pydantic_core_schema__(
-            cls,
-            source_type: Any,
-            handler: Callable[[Any], core_schema.CoreSchema],
-        ) -> core_schema.CoreSchema:
-            return core_schema.no_info_before_validator_function(
-                cls.validate,
-                core_schema.any_schema(),
-                serialization=core_schema.plain_serializer_function_ser_schema(
-                    cls._serialize, info_arg=False, return_schema=core_schema.str_schema(), when_used="json"
-                ),
-            )
+    @classmethod
+    def __get_pydantic_core_schema__(
+        cls,
+        source_type: Any,
+        handler: Callable[[Any], core_schema.CoreSchema],
+    ) -> core_schema.CoreSchema:
+        return core_schema.no_info_before_validator_function(
+            cls.validate,
+            core_schema.any_schema(),
+            serialization=core_schema.plain_serializer_function_ser_schema(
+                cls._serialize, info_arg=False, return_schema=core_schema.str_schema(), when_used="json"
+            ),
+        )
 
 
 def make_enum(name: str, values: Union[Dict[str, float], List], start=0):
