@@ -1,7 +1,6 @@
 from typing import Generic, Hashable, Sequence, Set, TypeVar
 
-import pydantic
-from packaging import version
+from pydantic import model_validator
 
 from .base import ContextBase, ResultBase
 
@@ -13,76 +12,40 @@ __all__ = (
 C = TypeVar("C", bound=Hashable)
 T = TypeVar("T")
 
-if version.parse(pydantic.__version__) < version.parse("2"):
-    from pydantic.generics import GenericModel
 
-    class GenericResult(ResultBase, GenericModel, Generic[T]):
-        """Holds anything."""
+class GenericResult(ResultBase, Generic[T]):
+    """Holds anything."""
 
-        value: T
+    value: T
 
-        @classmethod
-        def validate(cls, v, field=None):
-            if not isinstance(v, GenericResult) and not (isinstance(v, dict) and "value" in v):
-                v = {"value": v}
-            if isinstance(v, dict) and "value" in v:
-                if isinstance(v["value"], GenericContext):
-                    v["value"] = v["value"].value
-            return super(GenericResult, cls).validate(v)
+    @model_validator(mode="wrap")
+    def _validate_generic_result(cls, v, handler, info):
+        if isinstance(v, GenericResult) and not isinstance(v, cls):
+            v = {"value": v.value}
+        elif not isinstance(v, GenericResult) and not (isinstance(v, dict) and "value" in v):
+            v = {"value": v}
+        if isinstance(v, dict) and "value" in v:
+            if isinstance(v["value"], GenericContext):
+                v["value"] = v["value"].value
+        return handler(v)
 
-    class GenericContext(ContextBase, GenericModel, Generic[C]):
-        """Holds anything."""
 
-        value: C
+class GenericContext(ContextBase, Generic[C]):
+    """Holds anything."""
 
-        @classmethod
-        def validate(cls, v, field=None):
-            if not isinstance(v, GenericContext) and not (isinstance(v, dict) and "value" in v):
-                v = {"value": v}
-            if isinstance(v, dict) and "value" in v:
-                if isinstance(v["value"], GenericResult):
-                    v["value"] = v["value"].value
-                if isinstance(v["value"], Sequence) and not isinstance(v["value"], Hashable):
-                    v["value"] = tuple(v["value"])
-                if isinstance(v["value"], Set) and not isinstance(v["value"], Hashable):
-                    v["value"] = frozenset(v["value"])
-            return super(GenericContext, cls).validate(v)
+    value: C
 
-else:
-    from pydantic import model_validator
-
-    class GenericResult(ResultBase, Generic[T]):
-        """Holds anything."""
-
-        value: T
-
-        @model_validator(mode="wrap")
-        def _validate_generic_result(cls, v, handler, info):
-            if isinstance(v, GenericResult) and not isinstance(v, cls):
-                v = {"value": v.value}
-            elif not isinstance(v, GenericResult) and not (isinstance(v, dict) and "value" in v):
-                v = {"value": v}
-            if isinstance(v, dict) and "value" in v:
-                if isinstance(v["value"], GenericContext):
-                    v["value"] = v["value"].value
-            return handler(v)
-
-    class GenericContext(ContextBase, Generic[C]):
-        """Holds anything."""
-
-        value: C
-
-        @model_validator(mode="wrap")
-        def _validate_generic_context(cls, v, handler, info):
-            if isinstance(v, GenericContext) and not isinstance(v, cls):
-                v = {"value": v.value}
-            elif not isinstance(v, GenericContext) and not (isinstance(v, dict) and "value" in v):
-                v = {"value": v}
-            if isinstance(v, dict) and "value" in v:
-                if isinstance(v["value"], GenericResult):
-                    v["value"] = v["value"].value
-                if isinstance(v["value"], Sequence) and not isinstance(v["value"], Hashable):
-                    v["value"] = tuple(v["value"])
-                if isinstance(v["value"], Set) and not isinstance(v["value"], Hashable):
-                    v["value"] = frozenset(v["value"])
-            return handler(v)
+    @model_validator(mode="wrap")
+    def _validate_generic_context(cls, v, handler, info):
+        if isinstance(v, GenericContext) and not isinstance(v, cls):
+            v = {"value": v.value}
+        elif not isinstance(v, GenericContext) and not (isinstance(v, dict) and "value" in v):
+            v = {"value": v}
+        if isinstance(v, dict) and "value" in v:
+            if isinstance(v["value"], GenericResult):
+                v["value"] = v["value"].value
+            if isinstance(v["value"], Sequence) and not isinstance(v["value"], Hashable):
+                v["value"] = tuple(v["value"])
+            if isinstance(v["value"], Set) and not isinstance(v["value"], Hashable):
+                v["value"] = frozenset(v["value"])
+        return handler(v)
